requirements.txt
duckdb>=1.0.0
pyarrow>=15.0.0
polars>=1.7.0
pandas>=2.2.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
